functions/analyze_logs/requirements.txt
functions-framework==3.*
google-generativeai
google-cloud-logging